feat(bfdr): add additional layer of inheritance for ije natality
