+++ projects/BFDR/BFDR.xml
@@ -361,6 +361,9 @@
         <member name="P:BFDR.IJEBirth.Record">
             <summary>FHIR based vital record.</summary>
         </member>
+        <member name="P:BFDR.IJEBirth.NatalityRecord">
+            <summary>FHIR based natality record.</summary>
+        </member>
         <member name="M:BFDR.IJEBirth.ToBirthRecord">
             <summary>FHIR based vital record.</summary>
             Redundant due to ToRecord(), but kept for compatibility with IJEMortality which has this method for backwards compatibility.
@@ -1450,6 +1453,9 @@
         </member>
         <member name="P:BFDR.IJEFetalDeath.Record">
             <summary>FHIR based vital record.</summary>
+        </member>
+        <member name="P:BFDR.IJEFetalDeath.NatalityRecord">
+            <summary>FHIR based natality record.</summary>
         </member>
         <member name="M:BFDR.IJEFetalDeath.ToFetalDeathRecord">
             <summary>FHIR based vital record.</summary>
@@ -2557,6 +2563,34 @@
         </member>
         <member name="P:BFDR.IJEFetalDeath.BLANK2">
             <summary>Blank for Jurisdictional Use Only</summary>
+        </member>
+        <member name="T:BFDR.IJENatality">
+            <summary>Base class for all natality IJE converter classes</summary>
+        </member>
+        <member name="M:BFDR.IJENatality.#ctor">
+            <summary>Default constructor that creates a new, empty NatalityRecord.</summary>
+        </member>
+        <member name="M:BFDR.IJENatality.#ctor(BFDR.NatalityRecord,System.Boolean)">
+            <summary>Constructor that takes a <c>NatalityRecord</c>.</summary>
+        </member>
+        <member name="M:BFDR.IJENatality.#ctor(System.String,System.Boolean)">
+            <summary>Constructor that takes an IJE string and builds a corresponding internal <c>NatalityRecord</c>.</summary>
+        </member>
+        <member name="P:BFDR.IJENatality.NatalityRecord">
+            <summary>FHIR based natality record.</summary>
+        </member>
+        <member name="M:BFDR.IJENatality.Get_MotherRace(System.String)">
+            
+             <summary>Checks if the given race exists in the record for Mother.</summary>
+        </member>
+        <member name="M:BFDR.IJENatality.Set_MotherRace(System.String,System.String)">
+            <summary>Adds the given race to the record for Mother.</summary>
+        </member>
+        <member name="M:BFDR.IJENatality.Get_FatherRace(System.String)">
+            <summary>Checks if the given race exists in the record for Father.</summary>
+        </member>
+        <member name="M:BFDR.IJENatality.Set_FatherRace(System.String,System.String)">
+            <summary>Adds the given race to the record for Father.</summary>
         </member>
         <member name="T:BFDR.Mappings">
             <summary>Mappings between IJE and FHIR value sets</summary>
@@ -3811,6 +3845,17 @@
             <para>Console.WriteLine($"Ethnicity: {ExampleBirthRecord.MotherEthnicityLiteral["display"]}");</para>
             </example>
         </member>
+        <member name="P:BFDR.NatalityRecord.MotherRace">
+            <summary>Mother's Race values.</summary>
+            <value>the mother's race. A tuple, where the first value of the tuple is the display value, and the second is
+            the IJE code Y or N.</value>
+            <example>
+            <para>// Setter:</para>
+            <para>ExampleBirthRecord.MotherRace = {NvssRace.BlackOrAfricanAmerican, "Y"};</para>
+            <para>// Getter:</para>
+            <para>string boaa = ExampleBirthRecord.RaceBlackOfAfricanAmerican;</para>
+            </example>
+        </member>
         <member name="P:BFDR.NatalityRecord.FatherEthnicity1">
             <summary>Father's Ethnicity Hispanic Mexican.</summary>
             <value>the father's ethnicity. A Dictionary representing a code, containing the following key/value pairs:
@@ -3935,6 +3980,17 @@
             <para>ExampleBirthRecord.EthnicityLiteral = ethnicity;</para>
             <para>// Getter:</para>
             <para>Console.WriteLine($"Ethnicity: {ExampleBirthRecord.EthnicityLiteral["display"]}");</para>
+            </example>
+        </member>
+        <member name="P:BFDR.NatalityRecord.FatherRace">
+            <summary>Father's Race values.</summary>
+            <value>the father's race. A tuple, where the first value of the tuple is the display value, and the second is
+            the IJE code Y or N.</value>
+            <example>
+            <para>// Setter:</para>
+            <para>ExampleBirthRecord.FatherRace = {NvssRace.BlackOrAfricanAmerican, "Y"};</para>
+            <para>// Getter:</para>
+            <para>string boaa = ExampleBirthRecord.RaceBlackOfAfricanAmerican;</para>
             </example>
         </member>
         <member name="P:BFDR.NatalityRecord.DateOfLastLiveBirthDay">
