+++ projects/VitalRecord/VitalRecord_util.cs
@@ -1713,227 +1713,6 @@ namespace VR
             return null;
         }
 
-        /// <summary>Mother's Race values.</summary>
-        /// <value>the mother's race. A tuple, where the first value of the tuple is the display value, and the second is
-        /// the IJE code Y or N.</value>
-        /// <example>
-        /// <para>// Setter:</para>
-        /// <para>ExampleBirthRecord.MotherRace = {NvssRace.BlackOrAfricanAmerican, "Y"};</para>
-        /// <para>// Getter:</para>
-        /// <para>string boaa = ExampleBirthRecord.RaceBlackOfAfricanAmerican;</para>
-        /// </example>
-        [Property("MotherRace", Property.Types.TupleArr, "Race and Ethnicity Profiles", "Mother's Race", true, VR.IGURL.InputRaceAndEthnicity, true, 38)]
-        [FHIRPath("Bundle.entry.resource.where($this is Observation).where(code.coding.code='inputraceandethnicityMother')", "")]
-       // [FHIRSubject(FHIRSubject.Subject.Newborn)]
-        public Tuple<string, string>[] MotherRace
-        {
-            get
-            {
-                // filter the boolean race values
-                var booleanRaceCodes = NvssRace.GetBooleanRaceCodes();
-                List<string> raceCodes = booleanRaceCodes.Concat(NvssRace.GetLiteralRaceCodes()).ToList();
-
-                var races = new List<Tuple<string, string>>() { };
-
-                Observation obs = GetOrCreateObservation("inputraceandethnicityMother", CodeSystems.InputRaceAndEthnicityPerson, "Input Race and Ethnicity Person", VR.ProfileURL.InputRaceAndEthnicity, "inputraceandethnicityMother");
-
-                if (!obs.Component.Any())
-                {
-                    return races.ToArray();
-                }
-                foreach (string raceCode in raceCodes)
-                {
-                    Observation.ComponentComponent component = obs.Component.Where(c => c.Code.Coding[0].Code == raceCode).FirstOrDefault();
-                    if (component != null)
-                    {
-                        // convert boolean race codes to strings
-                        if (booleanRaceCodes.Contains(raceCode))
-                        {
-                            if (component.Value == null)
-                            {
-                                // If there is no value given, set the race to blank.
-                                var race = Tuple.Create(raceCode, "");
-                                races.Add(race);
-                                continue;
-                            }
-
-                            bool? raceBool = ((FhirBoolean)component.Value).Value;
-
-                            if (raceBool.Value)
-                            {
-                                var race = Tuple.Create(raceCode, "Y");
-                                races.Add(race);
-                            }
-                            else
-                            {
-                                var race = Tuple.Create(raceCode, "N");
-                                races.Add(race);
-                            }
-                        }
-                        else
-                        {
-                            // Ignore unless there's a value present
-                            if (component.Value != null)
-                            {
-                                var race = Tuple.Create(raceCode, component.Value.ToString());
-                                races.Add(race);
-                            }
-                        }
-
-                    }
-                }
-
-                return races.ToArray();
-            }
-            set
-            {
-                if (value.FirstOrDefault() == null) {
-                    return;
-                }
-                Observation obs = GetOrCreateObservation("inputraceandethnicityMother", CodeSystems.InputRaceAndEthnicityPerson, "Input Race and Ethnicity Person", VR.ProfileURL.InputRaceAndEthnicity, "inputraceandethnicityMother");
-                var booleanRaceCodes = NvssRace.GetBooleanRaceCodes();
-                var literalRaceCodes = NvssRace.GetLiteralRaceCodes();
-                foreach (Tuple<string, string> element in value)
-                {
-                    obs.Component.RemoveAll(c => c.Code.Coding[0].Code == element.Item1);
-                    Observation.ComponentComponent component = new Observation.ComponentComponent();
-                    String displayValue = NvssRace.GetDisplayValueForCode(element.Item1);
-                    component.Code = new CodeableConcept(CodeSystems.ComponentCodeVR, element.Item1, displayValue, null);
-                    if (booleanRaceCodes.Contains(element.Item1))
-                    {
-                        if (element.Item2 == "Y")
-                        {
-                            component.Value = new FhirBoolean(true);
-                        }
-                        else
-                        {
-                            component.Value = new FhirBoolean(false);
-                        }
-                    }
-                    else if (literalRaceCodes.Contains(element.Item1))
-                    {
-                        component.Value = new FhirString(element.Item2);
-                    }
-                    else
-                    {
-                        throw new ArgumentException("Invalid race literal code found: " + element.Item1 + " with value: " + element.Item2);
-                    }
-                    obs.Component.Add(component);
-                }
-
-            }
-        }
-
-        /// <summary>Father's Race values.</summary>
-        /// <value>the father's race. A tuple, where the first value of the tuple is the display value, and the second is
-        /// the IJE code Y or N.</value>
-        /// <example>
-        /// <para>// Setter:</para>
-        /// <para>ExampleBirthRecord.FatherRace = {NvssRace.BlackOrAfricanAmerican, "Y"};</para>
-        /// <para>// Getter:</para>
-        /// <para>string boaa = ExampleBirthRecord.RaceBlackOfAfricanAmerican;</para>
-        /// </example>
-        [Property("FatherRace", Property.Types.TupleArr, "Race and Ethnicity Profiles", "Father's Race", true, VR.IGURL.InputRaceAndEthnicity, true, 38)]
-        [FHIRPath("Bundle.entry.resource.where($this is Observation).where(code.coding.code='inputraceandethnicityFather')", "")]
-        // [FHIRSubject(FHIRSubject.Subject.Newborn)]
-        public Tuple<string, string>[] FatherRace
-        {
-            get
-            {
-                // filter the boolean race values
-                var booleanRaceCodes = NvssRace.GetBooleanRaceCodes();
-                List<string> raceCodes = booleanRaceCodes.Concat(NvssRace.GetLiteralRaceCodes()).ToList();
-
-                var races = new List<Tuple<string, string>>() { };
-
-                Observation obs = GetOrCreateObservation("inputraceandethnicityFather", CodeSystems.InputRaceAndEthnicityPerson, "Input Race and Ethnicity Person", VR.ProfileURL.InputRaceAndEthnicity, "inputraceandethnicityFather");
-
-                if (!obs.Component.Any())
-                {
-                    return races.ToArray();
-                }
-                foreach (string raceCode in raceCodes)
-                {
-                    Observation.ComponentComponent component = obs.Component.Where(c => c.Code.Coding[0].Code == raceCode).FirstOrDefault();
-                    if (component != null)
-                    {
-                        // convert boolean race codes to strings
-                        if (booleanRaceCodes.Contains(raceCode))
-                        {
-                            if (component.Value == null)
-                            {
-                                // If there is no value given, set the race to blank.
-                                var race = Tuple.Create(raceCode, "");
-                                races.Add(race);
-                                continue;
-                            }
-
-                            bool? raceBool = ((FhirBoolean)component.Value).Value;
-
-                            if (raceBool.Value)
-                            {
-                                var race = Tuple.Create(raceCode, "Y");
-                                races.Add(race);
-                            }
-                            else
-                            {
-                                var race = Tuple.Create(raceCode, "N");
-                                races.Add(race);
-                            }
-                        }
-                        else
-                        {
-                            // Ignore unless there's a value present
-                            if (component.Value != null)
-                            {
-                                var race = Tuple.Create(raceCode, component.Value.ToString());
-                                races.Add(race);
-                            }
-                        }
-
-                    }
-                }
-
-                return races.ToArray();
-            }
-            set
-            {
-                if (value.FirstOrDefault() == null) {
-                    return;
-                }
-                Observation obs = GetOrCreateObservation("inputraceandethnicityFather", CodeSystems.InputRaceAndEthnicityPerson, "Input Race and Ethnicity Person", VR.ProfileURL.InputRaceAndEthnicity, "inputraceandethnicityFather");
-                var booleanRaceCodes = NvssRace.GetBooleanRaceCodes();
-                var literalRaceCodes = NvssRace.GetLiteralRaceCodes();
-                foreach (Tuple<string, string> element in value)
-                {
-                    obs.Component.RemoveAll(c => c.Code.Coding[0].Code == element.Item1);
-                    Observation.ComponentComponent component = new Observation.ComponentComponent();
-                    String displayValue = NvssRace.GetDisplayValueForCode(element.Item1);
-                    component.Code = new CodeableConcept(CodeSystems.ComponentCodeVR, element.Item1, displayValue, null);
-                    if (booleanRaceCodes.Contains(element.Item1))
-                    {
-                        if (element.Item2 == "Y")
-                        {
-                            component.Value = new FhirBoolean(true);
-                        }
-                        else
-                        {
-                            component.Value = new FhirBoolean(false);
-                        }
-                    }
-                    else if (literalRaceCodes.Contains(element.Item1))
-                    {
-                        component.Value = new FhirString(element.Item2);
-                    }
-                    else
-                    {
-                        throw new ArgumentException("Invalid race literal code found: " + element.Item1 + " with value: " + element.Item2);
-                    }
-                    obs.Component.Add(component);
-                }
-            }
-        }
-
         // /// <summary>Check to make sure the given profile contains the given resource.</summary>
         // protected static bool MatchesProfile(string resource, string profile)
         // {
